Move keyphraseLenghtFactor.js to assessment helpers
