+++ packages/yoastseo/src/scoring/helpers/assessments/keyphraseLengthFactor.js
